Work around exception bug in tests
When linking against LLVM12 the codegen tests fail:

```
0  0x00005555577275ae in classify_object_over_fdes ()
1  0x0000555557727a48 in search_object ()
2  0x0000555557728436 in _Unwind_Find_FDE ()
3  0x00005555577246b3 in uw_frame_state_for ()
4  0x0000555557725890 in uw_init_context_1 ()
5  0x00005555577260ae in _Unwind_RaiseException ()
6  0x0000555557685d57 in __cxa_throw ()
7  0x0000555555a085f0 in std::__throw_invalid_argument(char const*) [clone .cold.2] ()
8  0x0000555555ca8fc5 in long long __gnu_cxx::__stoa<long long, long long, char, int>(long long (*)(char const*, char**, int), char const*, char const*, unsigned long*, int) ()
9  0x0000555555ca81d1 in std::__cxx11::stoll(std::__cxx11::basic_string<char, std::char_traits<char>, std::allocator<char> > const&, unsigned long*, int) ()
10 0x0000555555d259d9 in bpftrace::ast::AttachPointParser::uprobe_parser(bool, bool) ()
```

It seems like exception handling somehow breaks in these functions. It
only happens when these functions are called multiple times, so it only
happens in the test suite. I haven't seen these pop up in normal usage
yet.

This work around is hacky and not something to be proud of, but it
allows us to switch to LLVM 12 so we can do it for now.