parser: move attachpoint detection to the lexer
Due to its dynamic nature we created a new parser for attach points but
it's input still comes from the lexer/parser step. The current form of
tokenizing it in the lexer and trying to combine it again in the parser
is tedious and not as flexible as it could be. We have to go through
some effort to ensure both the lexer and the parser allow the sequence
we need.

This takes a different approach. All the logic has been moved to the
lexer, the parser just receives a single "ATTACH_POINT". To help
detection the parser signals its expected state to the lexer through the
'yyps' state variable. This is not strictly needed as we don't allow
bare words, except for builtins, so there shouldn't be a parsing
conflict but the state helps with the error message.

The listing syntax is more flexible than attach syntax is, e.g.
`*nanosleep*` is allowed in listing which is hard to detect in the
lexer. To work around this a '\a' gets inserted at the start of the
list query, which sends the parser into specific state for listing. This
is a bit ugly but should be unnoticeable in most situations.

The support for wildcard probe types in (non listing) programs has been
dropped due to this. But I don't see a valid usecase for it